requirements.txt
numpy
scikit-learn
matplotlib
streamlit
openai
python-dotenv